renamed slug service and moved dependency
