+++ services/package.json
@@ -7,8 +7,7 @@
 		"kysely": "^0.27.3",
 		"pino": "^9.0.0",
 		"pino-pretty": "^11.0.0",
-		"zod": "^3.23.5",
-		"slugify": "^1.6.6"
+		"zod": "^3.23.5"
 	},
 	"devDependencies": {
 		"vite": "^5.2.11",
